:recycle: 검색 기록 창 smart-dumb 분리 리팩토링
